serverside/collect_email/requirements.txt
requests
google-cloud-error-reporting